Create Certainly! Here's an example PHP code that takes input from a file, downloads a remote file, and saves it locally:
Make sure to update the $inputFilePath and $outputFilePath variables with the appropriate paths for your case. The input file should contain the URL of the remote file you want to download, with no additional content or whitespace.

When you run this PHP script, it will read the remote file URL from the input file, download the file, and save it locally with the specified output file path. You will see appropriate messages indicating whether the process was successful or if any errors occurred.